Replace the TiddlySpace backstage with a stub tsbar
This is just to experiment with the placement in various places.  At the moment a single JS tiddler is being created in the common bag, derived from a a stub file in the clean-slate branch of the tsbar repository.

For tiddlywiki, there is a tweak to the CSS to push the TiddlyWiki backstage below the tsbar.